rename: defs.rs --> common.rs
